agent/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.64.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
distro==1.9.0
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-anthropic==0.3.19
langchain-core==0.3.74
langchain-tavily==0.2.11
langchain-text-splitters==0.3.9
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-checkpoint-sqlite==2.0.11
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.3
langsmith==0.4.16
multidict==6.6.4
orjson==3.11.2
ormsgpack==1.10.0
packaging==25.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.0.0
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
schedule==1.2.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlite-vec==0.1.6
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.24.0
